feat: refactor records handling by extracting routes into separate file and simplifying endpoint logic
